requirements.txt
aiogram>=3.0.0b8
aiohttp
python-dotenv